+++ MarkdownMonster/Windows/Git/GitCommitModel.cs
@@ -144,6 +144,24 @@ namespace MarkdownMonster.Windows
         private string _Branch;
 
 
+        public List<string> LocalBranches
+        {
+            get
+            {
+                if (GitHelper.Repository?.Branches == null)
+                    return null;
+
+                var list = GitHelper.Repository.Branches
+                    .Where(b => b.IsRemote == false)
+                    .Select( b=> b.FriendlyName)                    
+                    .ToList();
+
+                list.Insert(0, "<Create new Branch...>");
+                return list;
+            }            
+        }
+
+
         public string Remote
         {
             get { return _Remote; }
@@ -168,7 +186,7 @@ namespace MarkdownMonster.Windows
             }
         }
         private ObservableCollection<RepositoryStatusItem> _repositoryStatusItems;
-
+        
         public AppModel AppModel { get; set; }
 
         public MainWindow Window { get; set; }
@@ -191,6 +209,9 @@ namespace MarkdownMonster.Windows
             }
         }
 
+        
+
+
         #region Load
 
         public GitCommitModel(string fileOrFolder, bool commitRepository = false)
@@ -208,9 +229,11 @@ namespace MarkdownMonster.Windows
                 var userEmail = GitHelper.GetGitNameAndEmailFromGitConfig();
                 GitUsername = userEmail[0];
                 GitEmail = userEmail[1];
-            }
+            }            
 
             ShowUserInfo = string.IsNullOrEmpty(GitUsername);
+
+            
         }
         #endregion